refactor(feed page): create ride search hook, simplify TopBar
